Add ask_fandom/parser.py
